Clear project without storyboards
