+++ DiscountsAndPromotionsApp/Application/SceneDelegate.swift
@@ -1,0 +1,13 @@
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let scene = (scene as? UIWindowScene) else { return }
+        let window = UIWindow(windowScene: scene)
+        window.rootViewController = MainViewController()
+        self.window = window
+        window.makeKeyAndVisible()
+    }
+}
