+++ DiscountsAndPromotionsApp/MainViewController.swift
@@ -1,0 +1,10 @@
+
+import UIKit
+
+class MainViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .red
+    }
+}
