+++ DiscountsAndPromotionsApp/Resources/Assets.xcassets/Contents.json
