--- DiscountsAndPromotionsApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  DiscountsAndPromotionsApp
-//
-//  Created by Павел Афанасьев on 13.11.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
